# services/cell-api/internal/api/openapi_gen.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "description": "Personal access token (con_pat_…) or service-account token (con_sa_…).",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "Sluicio-Session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Sluicio cell-api. Authenticate with a session cookie (browser) or a Bearer token — a personal access token or a service-account token (see Settings → Service accounts). Generated from the route table; do not edit by hand.",
    "title": "Sluicio API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/integrations": {
      "get": {
        "operationId": "get_api_v1_integrations",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "List integrations",
        "tags": [
          "integrations"
        ]
      },
      "post": {
        "operationId": "post_api_v1_integrations",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create integration",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/v1/maps": {
      "get": {
        "operationId": "get_api_v1_maps",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "List maps",
        "tags": [
          "maps"
        ]
      },
      "post": {
        "operationId": "post_api_v1_maps",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create map",
        "tags": [
          "maps"
        ]
      }
    },
...
    "/api/v1/metadata-fields": {
      "get": {
        "operationId": "get_api_v1_metadata_fields",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "List metadata-fields",
        "tags": [
          "metadata-fields"
        ]
      },
      "post": {
        "operationId": "post_api_v1_metadata_fields",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create metadata-field",
        "tags": [
          "metadata-fields"
        ]
      }
    },
...
    "/api/v1/schemas": {
      "get": {
        "operationId": "get_api_v1_schemas",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "List schemas",
        "tags": [
          "schemas"
        ]
      },
      "post": {
        "operationId": "post_api_v1_schemas",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create schema",
        "tags": [
          "schemas"
        ]
      }
    },
...
    "/api/v1/service-facets": {
      "get": {
        "operationId": "get_api_v1_service_facets",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "List service-facets",
        "tags": [
          "service-facets"
        ]
      },
      "post": {
        "operationId": "post_api_v1_service_facets",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create service-facet",
        "tags": [
          "service-facets"
        ]
      }
    },
...
    "/api/v1/services": {
      "get": {
        "operationId": "get_api_v1_services",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "List services",
        "tags": [
          "services"
        ]
      }
    },
...
    "/api/v1/tags": {
      "get": {
        "operationId": "get_api_v1_tags",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "List tags",
        "tags": [
          "tags"
        ]
      },
      "post": {
        "operationId": "post_api_v1_tags",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create tag",
        "tags": [
          "tags"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
...
}